Fix barotropic streamfunction calculation
Instead of computing the BSF using a least-squares solve based
on the transport between vertices, the new approach inverts
a Poisson equation for the BSF based on the (vertically integrated)
vorticity.

I have proven that this produces more accurate results.  The
vertically integrated vorticity computed from the streamfunciton
is within machine precision of the original vertically integrated
vorticity.  The velocity produced by the streamfunciton has errors
that are on the order of 1e-3 compared with the original vertically
integrated velocity (likely caused by the field not being perfectly
divergence free).